--- web/src/features/history/WeeklyHistory.tsx
@@ -1,9 +1,0 @@
-const WeeklyHistory = () => {
-    return (
-        <div>
-            list of daily stuff
-        </div>
-    )
-}
-
-export { WeeklyHistory };
